feat: add withtheme for alteast until hooks work
