+++ addons/ondevice-notes/src/register.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import addons from '@storybook/addons';
-import { Notes } from './components/Notes';
+import Notes from './components/Notes';
 
 export const PARAM_KEY = `notes`;
