+++ app/react-native-server/package.json
@@ -41,7 +41,7 @@
     "core-js": "^3.0.1",
     "global": "^4.3.2",
     "react": "16.13.1",
-    "react-dom": "^16.8.3",
+    "react-dom": "^16.13.1",
     "uuid": "^3.3.2",
     "webpack": "^4.33.0",
     "ws": "^7.1.2"
