+++ lib/theme/package.json
@@ -24,11 +24,7 @@
   "scripts": {
     "prepare": "node ../../scripts/prepare.js"
   },
-  "dependencies": {
-    "@emotion/native": "^10.0.27",
-    "@emotion/styled-base": "^10.0.31",
-    "emotion-theming": "^10.0.19"
-  },
+  "dependencies": {},
   "peerDependencies": {
     "react": "*",
     "react-native": "*"
