+++ lib/theme/src/index.ts
@@ -1,4 +1,4 @@
 export { useTheme } from './useTheme';
 export { theme, Theme } from './theme';
-export { ThemeProvider } from './ThemeProvider';
+export { ThemeProvider, withTheme } from './ThemeContext';
 export { styled, ThemedStyle, createStyled } from './Styled';
