+++ lib/theme/src/useTheme.ts
@@ -1,5 +1,5 @@
 import React, { useContext } from 'react';
-import { ThemeContext } from './ThemeProvider';
+import { ThemeContext } from './ThemeContext';
 
 export function useTheme() {
   const theme = useContext(ThemeContext);
