+++ package.json
@@ -126,7 +126,6 @@
     "@babel/preset-react": "^7.0.0",
     "@babel/preset-typescript": "^7.3.3",
     "@babel/runtime": "^7.7.2",
-    "@emotion/snapshot-serializer": "^0.8.2",
     "@jest/transform": "^25.1.0",
     "@storybook/eslint-config-storybook": "^2.0.0",
     "@storybook/linter-config": "^2.0.0",
@@ -148,10 +147,8 @@
     "babel-loader": "^8.0.5",
     "babel-plugin-add-react-displayname": "^0.0.5",
     "babel-plugin-dynamic-import-node": "^2.2.0",
-    "babel-plugin-emotion": "^10.0.20",
     "babel-plugin-macros": "^2.7.0",
     "babel-plugin-require-context-hook": "^1.0.0",
-    "babel-preset-vue": "^2.0.2",
     "chalk": "^3.0.0",
     "codecov": "^3.5.0",
     "codelyzer": "^5.0.0",
@@ -177,7 +174,6 @@
     "jest-cli": "^24.8.0",
     "jest-config": "^24.8.0",
     "jest-diff": "^24.8.0",
-    "jest-emotion": "^10.0.17",
     "jest-environment-jsdom": "^24.8.0",
     "jest-environment-jsdom-thirteen": "^1.0.1",
     "jest-enzyme": "^7.0.2",
@@ -185,7 +181,6 @@
     "jest-jasmine2": "^24.8.0",
     "jest-matcher-utils": "^24.8.0",
     "jest-raw-loader": "^1.0.1",
-    "jest-vue-preprocessor": "^1.5.0",
     "jest-watch-typeahead": "^0.4.0",
     "lerna": "^3.19.0",
     "lint-staged": "^9.4.1",
